refactor: unify git query host seams
