+++ apps/desktop/src/features/agent-studio-git/use-agent-studio-diff-data.test.tsx
@@ -80,12 +80,6 @@ const gitGetWorktreeStatusSummaryMock = mock(
 mock.module("@/lib/host-client", () => ({
   hostClient: {
     runsList: runsListMock,
-  },
-}));
-
-mock.module("@/state/operations/host", () => ({
-  host: {
-    runsList: runsListMock,
     gitGetWorktreeStatus: gitGetWorktreeStatusMock,
     gitGetWorktreeStatusSummary: gitGetWorktreeStatusSummaryMock,
   },
