+++ apps/desktop/src/features/agent-studio-git/use-agent-studio-diff-data.ts
@@ -1,11 +1,11 @@
 import { useCallback, useMemo, useState } from "react";
-import type { DiffDataState } from "@/features/agent-studio-git";
 import { canonicalTargetBranch } from "@/lib/target-branch";
 import type { UseAgentStudioDiffDataInput } from "./agent-studio-diff-data-model";
+import type { DiffDataState } from "./contracts";
 import { useAgentStudioDiffController } from "./use-agent-studio-diff-controller";
 import { useAgentStudioWorktreeResolution } from "./use-agent-studio-worktree-resolution";
 
-export type { DiffDataState, DiffScope } from "@/features/agent-studio-git";
+export type { DiffDataState, DiffScope } from "./contracts";
 
 export function useAgentStudioDiffData({
   repoPath,
